% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed0.R
\name{ed0}
\alias{ed0}
\title{Compute in-air downward irradiance (Ed0)}
\usage{
ed0(yday, hour, lat, lon, tcl, o3, cf, albedo = 0.05)
}
\arguments{
\item{yday}{Numeric. Day of the year.}

\item{hour}{Numeric. Hour of the day.}

\item{lat}{Numeric. Latitude.}

\item{lon}{Numeric. Longitude.}

\item{tcl}{Numeric. Cloud optical thickness.}

\item{o3}{Numeric. Ozone.}

\item{cf}{Numeric. Cloud fraction.}

\item{albedo}{Numeric. Surface albedo.}
}
\value{
Numeric vector containing downward irradiance spectra from 290 nm to 700 nm at 5 nm resolution.
}
\description{
Compute in-air downward irradiance (Ed0)
}
\examples{
ed0(100, 12, 67.47973, -63.78953, 3, 330, 1, 0.05)
}
