% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edpar.R
\name{edpar}
\alias{edpar}
\title{Compute in-air downward PAR}
\usage{
edpar(yday, hour, lat, lon, tcl, o3, cf, albedo = 0.05)
}
\arguments{
\item{yday}{Numeric. Day of the year.}

\item{hour}{Numeric. Hour of the day.}

\item{lat}{Numeric. Latitude.}

\item{lon}{Numeric. Longitude.}

\item{tcl}{Numeric. Cloud optical thickness.}

\item{o3}{Numeric. Ozone.}

\item{cf}{Numeric. Cloud fraction.}

\item{albedo}{Numeric. Surface albedo.}
}
\value{
Numeric. PAR calculated bwetween 400 and 700 nm expressed in umol photons m-2 s-1.
}
\description{
Compute in-air downward PAR
}
\examples{
edpar(100, 12, 67.47973, -63.78953, 3, 330, 1, 0.05)
}
