% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed0.R
\name{ed0}
\alias{ed0}
\title{Compute in-air downward irradiance (Ed0)}
\usage{
ed0(yday, hour, lat, lon, tcl, o3, cf, albedo = 0.05,
  lut_type = "ed0-")
}
\arguments{
\item{yday}{Numeric. Day of the year.}

\item{hour}{Numeric. Hour of the day in UTC (0 to 24).}

\item{lat}{Numeric. Latitude.}

\item{lon}{Numeric. Longitude.}

\item{tcl}{Numeric. Cloud optical thickness (0 to 64.0).}

\item{o3}{Numeric. Ozone concentration (100 to 550 DU).}

\item{cf}{Numeric. Cloud fraction (0 to 1.0).}

\item{albedo}{Numeric. Surface albedo (0 to 1.0).}

\item{lut_type}{Character. Either "ed0+" for in-air irradiance or "ed0-" for underwater irradiance.}
}
\value{
Numeric vector containing downward irradiance spectra from 290 nm to 700 nm at 5 nm resolution.
}
\description{
Compute in-air downward irradiance (Ed0)
}
\examples{
ed0(100, 12, 67.47973, -63.78953, 3, 330, 1, 0.05, "ed0+")
}
